french_employment/requirements.txt
dash
pandas
folium
geopandas
numpy
branca
plotly.express
dash_html_components
dash_core_components

